+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Changelog
-- try fix compatible with include LDL mods
-- `Forgified Fabric API` compat
+- try fix ItemLightSources register